# superserver/internal/controller/label/create/schema.go
package create

import (
	"github.com/gin-gonic/gin"
	"superserver/internal/controller"
)

type Controller struct {
	*controller.Controller
}

func NewController(ctx *gin.Context) *Controller {
	return &Controller{controller.New(ctx, new(Params))}
}

type Params struct {
	Source int32  `json:"source"`
	Name   string `json:"name"`
}

type Reply struct{}